requirements.txt
appnope==0.1.0
APScheduler==3.6.3
arrow==0.17.0
asgiref==3.4.1
astroid==2.4.2
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
chromedriver==2.24.1
cloudinary==1.26.0
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==35.0.0
decorator==5.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.46.0
django-bootstrap-v5==1.0.6
django-bootstrap4==3.0.1
django-bootstrap5==2.1.2
django-cors-headers==3.10.0
django-crontab==0.7.1
django-heroku==0.3.1
django-myapp==0.1
django-phonenumber-field==6.0.0
django-registration-redux==2.9
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
future==0.18.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
importlib-metadata==3.10.1
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
myapp==0.1.dev0
oauthlib==3.1.1
outcome==1.1.0
parso==0.7.1
pexpect==4.8.0
phonenumbers==8.12.37
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2==2.9.1
ptyprocess==0.6.0
pycparser==2.21
Pygments==2.7.2
PyJWT==2.3.0
pylint==2.6.0
pyOpenSSL==21.0.0
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.5
python-dotenv==0.19.1
python-twitter==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
schedule==1.1.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.1
traitlets==5.0.5
trio==0.19.0
trio-websocket==0.9.2
tweepy==4.3.0
tzlocal==2.1
urllib3==1.26.7
wcwidth==0.2.5
webdriver-manager==3.5.2
whitenoise==5.3.0
wrapt==1.12.1
wsproto==1.0.0
zipp==3.6.0
